Big update of front cam detector
Moved all the code from my separate repo to the main MAV course repo
Reorganised it a bit to separate dataset preparation from the training files.
Included the dockerfile used to convert ONNX models to C.
Added my name to the files (hopefully didn't miss any)
Tested it with a subset of the data to verify the training still runs well.